Title: 합과 차의 나눗셈, Time: 170ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240530/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -6,7 +6,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int a = sc.nextInt();
         int b = sc.nextInt();
-        double as = (a+b)/(a-b);
+        double as = (double)(a+b)/(a-b);
         System.out.printf("%.2f",as);
     }
 }
